// src/foundation/validation/application/models/JsonSchemaId.ts
export enum JsonSchemaId {
  cardFindNumberEnumQueryFilter = 'https://tgc-catalog/v1/cards/card-find-number-enum-query-filter',
  cardFindStringEnumQueryFilter = 'https://tgc-catalog/v1/cards/card-find-string-enum-query-filter',
  cardFindOptionalStringEnumQueryFilter = 'https://tgc-catalog/v1/cards/card-find-optional-string-enum-query-filter',
  cardFindQuery = 'https://tgc-catalog/v1/cards/card-find-query',
  cardFindQueryAndFilter = 'https://tgc-catalog/v1/cards/card-find-query-and-filter',
  cardFindQueryListFilter = 'https://tgc-catalog/v1/cards/card-find-query-list-filter',
  cardFindQueryNullFilter = 'https://tgc-catalog/v1/cards/card-find-query-null-filter',
  cardFindQueryOrFilter = 'https://tgc-catalog/v1/cards/card-find-query-or-filter',
  cardFindQueryRangeFilter = 'https://tgc-catalog/v1/cards/card-find-query-range-filter',
  cardFindQueryTextFilter = 'https://tgc-catalog/v1/cards/card-find-query-text-filter',
  cardFindQueryValueFilter = 'https://tgc-catalog/v1/cards/card-find-query-value-filter',
}